requirements.txt
python-novaclient==2.23.2
python-glanceclient==2.5
python-cinderclient==1.6
python-neutronclient==2.3.4
urllib3[secure]
